Style adjustments
Adding a reset to theme default item to the revisions list.

Adding a dependencies to the `useMemo` so that it knows to update when the number of revisions change.

This is so we can ensure we add the right metadata in the right order, e.g., `isLatest`

Extracted hooks and components from screen revisions monster file
