+++ packages/edit-site/src/components/global-styles/style.scss
@@ -192,17 +192,27 @@
 
 .edit-site-global-styles-screen-revisions__revision-item {
 	position: relative;
-	padding-left: $grid-unit-15;
+	padding: $grid-unit-10 0 $grid-unit-10 $grid-unit-15;
+
+	&:first-child {
+		padding-top: 0;
+	}
+
+	&:last-child {
+		padding-bottom: 0;
+	}
+
 	&::before {
 		background: $gray-300;
 		border-radius: 50%;
 		content: "\a";
 		display: inline-block;
-		height: 6px;
-		width: 6px;
+		height: $grid-unit-10;
+		width: $grid-unit-10;
 		position: absolute;
 		top: 50%;
-		left: -3px;
+		left: 0;
+		transform: translate(-50%, -50%);
 	}
 	&.is-current::before {
 		background: var(--wp-components-color-accent, var(--wp-admin-theme-color, #007cba));
@@ -213,13 +223,27 @@
 	width: 100%;
 	height: auto;
 	display: block;
-	min-height: $grid-unit-80;
+	padding: $grid-unit-10 $grid-unit-15;
+
+	&:hover {
+		background: rgba(var(--wp-admin-theme-color--rgb), 0.04);
+
+		.edit-site-global-styles-screen-revisions__date {
+			color: var(--wp-admin-theme-color);
+		}
+	}
 }
 
-.is-current > .edit-site-global-styles-screen-revisions__revision-button {
-	color: var(--wp-components-color-accent, var(--wp-admin-theme-color, #007cba));
-	opacity: 1;
-	background-color: #f1faff;
+.is-current {
+	.edit-site-global-styles-screen-revisions__revision-button {
+		color: var(--wp-components-color-accent, var(--wp-admin-theme-color, #007cba));
+		opacity: 1;
+		background: rgba(var(--wp-admin-theme-color--rgb), 0.04);
+	}
+
+	.edit-site-global-styles-screen-revisions__date {
+		color: var(--wp-admin-theme-color);
+	}
 }
 
 .edit-site-global-styles-screen-revisions__button {
@@ -228,33 +252,23 @@
 }
 
 .edit-site-global-styles-screen-revisions__description {
-	display: grid;
-	grid-template-columns: 1fr max-content;
-	grid-template-rows: 1fr 1fr;
-	gap: $grid-unit-05 $grid-unit-10;
-	grid-auto-flow: row;
-	align-items: start;
-	justify-items: start;
-	grid-template-areas:
-		". edit-site-global-styles-screen-revisions__avatar"
-		". edit-site-global-styles-screen-revisions__avatar";
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	gap: $grid-unit-10;
+}
 
-	.edit-site-global-styles-screen-revisions__avatar {
-		grid-area: edit-site-global-styles-screen-revisions__avatar;
-		width: 24px;
-		height: 24px;
-		align-self: center;
-		overflow: hidden;
-		border-radius: 50%;
-		img {
-			display: inline;
-			margin: 0 auto;
-			height: 100%;
-			width: auto;
-		}
-	}
-	.edit-site-global-styles-screen-revisions__date {
-		font-size: 11px;
+.edit-site-global-styles-screen-revisions__date {
+	color: $gray-700;
+	display: flex;
+	justify-content: space-between;
+	width: 100%;
+	align-items: center;
+
+	img {
+		width: $grid-unit-20;
+		height: $grid-unit-20;
+		border-radius: 100%;
 	}
 }
